Oracle DDL Done.
git-svn-id: svn://192.168.1.10/public/CoffeeMud@24338 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/core/database/DDLGenerator.java
@@ -83,6 +83,13 @@ public class DDLGenerator
 			}
 		}
 		R.close();
+		if (this.productName.contains("oracle")) 
+		{
+			typeMappings.put("INT", "INTEGER");
+			typeMappings.put("LONG", "NUMBER(20,0)");
+			typeMappings.put("VARCHAR", "VARCHAR2");
+			typeMappings.put("TEXT", "CLOB");
+		}
 	}
 
 	/**
@@ -96,7 +103,7 @@ public class DDLGenerator
 	{
 		final String quote=metaData.getIdentifierQuoteString();
 		final StringBuilder sql = new StringBuilder("DROP INDEX ").append(quote).append(indexName).append(quote);
-		if((!productName.contains("derby"))&&(!productName.contains("hsql")))
+		if((!productName.contains("derby"))&&(!productName.contains("hsql"))&&(!productName.contains("oracle")))
 			 sql.append(" ON ").append(quote).append(tableName).append(quote);
 		return sql.toString();
 	}
@@ -223,7 +230,7 @@ public class DDLGenerator
 		}
 		colRs.close();
 
-		final String effectiveNewType = (newType != null) ? mapPortableType(newType) : oldType;
+		final String effectiveNewType = (newType != null) ? mapPortableToLocalType(newType) : oldType;
 		final Integer effectiveNewSizeInt = (newSize != null) ? newSize : Integer.valueOf(oldSize);
 		final int effectiveNewSize = effectiveNewSizeInt.intValue();
 		final Boolean effectiveNewNullableBool = (newNullable != null) ? newNullable : Boolean.valueOf(oldNullable);
@@ -393,7 +400,7 @@ public class DDLGenerator
 		}
 		colRs.close();
 
-		final String effectiveNewType = (newType != null) ? mapPortableType(newType) : oldType;
+		final String effectiveNewType = (newType != null) ? mapPortableToLocalType(newType) : oldType;
 		final Integer effectiveNewSizeInt = (newSize != null) ? newSize : Integer.valueOf(oldSize);
 		final int effectiveNewSize = effectiveNewSizeInt.intValue();
 		final Boolean effectiveNewNullableBool = (newNullable != null) ? newNullable : Boolean.valueOf(oldNullable);
@@ -861,12 +868,185 @@ public class DDLGenerator
 	 * Generates MODIFY COLUMN SQL with new size and nullability
 	 * Note: Modifying nullability or size may have restrictions in some DBs (e.g., can't reduce size if data exists)
 	 */
+	@SuppressWarnings("unchecked")
+	public List<String> getModifyColumnSQLOracle(final String actualTable, final String columnName, final String newType, final Integer newSize, final Boolean newNullable) throws SQLException 
+	{
+		final List<String> sqls = new ArrayList<String>();
+		final String quote = metaData.getIdentifierQuoteString();
+		final String tempColumnName = columnName + "_TEMP";
+
+		String oldType = "";
+		String oldPortableType = "";
+		int oldSize = -1;
+		boolean oldNullable = true;
+		final ResultSet colRs = metaData.getColumns(null, null, actualTable, columnName);
+		if(colRs.next()) 
+		{
+			oldType = colRs.getString("TYPE_NAME").toUpperCase();
+			oldSize = colRs.getInt("COLUMN_SIZE");
+			oldNullable = "YES".equals(colRs.getString("IS_NULLABLE"));
+			final int oldDataType = colRs.getInt("DATA_TYPE");
+			oldPortableType = sqlTypeToPortable.getOrDefault(Integer.valueOf(oldDataType), oldType);
+		}
+		colRs.close();
+
+		final String effectivePortableType = (newType != null) ? newType : oldPortableType;
+		final String effectiveNewType = (newType != null) ? mapPortableToLocalType(newType) : oldType;
+		final Integer effectiveNewSizeInt = (newSize != null) ? newSize : Integer.valueOf(oldSize);
+		final int effectiveNewSize = effectiveNewSizeInt.intValue();
+		final Boolean effectiveNewNullableBool = (newNullable != null) ? newNullable : Boolean.valueOf(oldNullable);
+		final boolean effectiveNewNullable = effectiveNewNullableBool.booleanValue();
+		final List<String> pkCols = new ArrayList<String>();
+		String pkName = null;
+		final ResultSet pkRs = metaData.getPrimaryKeys(null, null, actualTable);
+		while (pkRs.next()) 
+		{
+			pkCols.add(pkRs.getString("COLUMN_NAME").toUpperCase());
+			if(pkName == null) pkName = pkRs.getString("PK_NAME");
+		}
+		pkRs.close();
+		final boolean isPkColumn = pkCols.contains(columnName.toUpperCase());
+		final Map<String, Map<String, Object>> indexes = new LinkedHashMap<String, Map<String, Object>>();
+		final ResultSet idxRs = metaData.getIndexInfo(null, null, actualTable, false, false);
+		while (idxRs.next()) 
+		{
+			final short type = idxRs.getShort("TYPE");
+			if(type == DatabaseMetaData.tableIndexStatistic) 
+				continue;
+			final String indexName = idxRs.getString("INDEX_NAME");
+			if((indexName == null)
+			||((pkName != null)
+				&&(indexName.equals(pkName) || indexName.startsWith("SYS_IDX_")))) 
+					continue;
+			final boolean nonUnique = idxRs.getBoolean("NON_UNIQUE");
+			final String colName = idxRs.getString("COLUMN_NAME");
+			if(colName == null) 
+				continue;
+			final short ordinal = idxRs.getShort("ORDINAL_POSITION");
+			final Map<String, Object> index = indexes.computeIfAbsent(indexName, k -> 
+			{
+				final Map<String, Object> m = new HashMap<String, Object>();
+				m.put("unique", Boolean.valueOf(!nonUnique));
+				m.put("columns", new ArrayList<Map<String, Object>>());
+				return m;
+			});
+			final Map<String, Object> colMap = new HashMap<String, Object>();
+			colMap.put("pos", Short.valueOf(ordinal));
+			colMap.put("col", colName.toUpperCase());
+			((List<Map<String, Object>>) index.get("columns")).add(colMap);
+		}
+		idxRs.close();
+		final List<Map<String, Object>> depIndexes = new ArrayList<Map<String, Object>>();
+		for (final Map.Entry<String, Map<String, Object>> entry : indexes.entrySet()) 
+		{
+			final String indexName = entry.getKey();
+			final Map<String, Object> index = entry.getValue();
+			final List<Map<String, Object>> cols = (List<Map<String, Object>>) index.get("columns");
+			boolean depends = false;
+			for (final Map<String, Object> c : cols) 
+			{
+				if(((String) c.get("col")).equals(columnName.toUpperCase())) 
+				{
+					depends = true;
+					break;
+				}
+			}
+			if(depends) 
+			{
+				final Map<String, Object> dep = new HashMap<String, Object>();
+				dep.put("name", indexName);
+				dep.put("unique", index.get("unique"));
+				cols.sort((a, b) -> ((Short) a.get("pos")).compareTo((Short) b.get("pos")));
+				final List<String> colList = new ArrayList<String>();
+				for (final Map<String, Object> c : cols) 
+					colList.add((String) c.get("col"));
+				dep.put("columns", colList);
+				depIndexes.add(dep);
+			}
+		}
+		if(isPkColumn) 
+		{
+			String dropSql;
+			if(pkName != null)
+				dropSql = "ALTER TABLE " + quote + actualTable + quote + " DROP CONSTRAINT " + quote + pkName + quote;
+			else
+				dropSql = "ALTER TABLE " + quote + actualTable + quote + " DROP PRIMARY KEY";
+			sqls.add(dropSql);
+		}
+		for (final Map<String, Object> idx : depIndexes) 
+		{
+			final String indexName = (String) idx.get("name");
+			sqls.add(getDropIndexSQL(indexName, actualTable));
+		}
+		final String addTempSql = "ALTER TABLE " + quote + actualTable + quote + " ADD " + quote + tempColumnName + quote + " " 
+								  + getDataTypeSQL(effectivePortableType, effectiveNewSizeInt);
+		sqls.add(addTempSql);
+		String copyExpr = quote + columnName + quote;
+		if((effectiveNewSize < oldSize) && effectiveNewType.contains("VARCHAR"))
+			copyExpr = "SUBSTR(" + copyExpr + ", 1, " + effectiveNewSize + ")";
+		else if(effectiveNewType.equals("CLOB") && oldType.contains("CHAR"))
+			copyExpr = "TO_CLOB(" + copyExpr + ")";
+		final String copySql = "UPDATE " + quote + actualTable + quote + " SET " + quote + tempColumnName + quote + " = " + copyExpr;
+		sqls.add(copySql);
+		if(!effectiveNewNullable) 
+		{
+			final String replaceNullsSql = "UPDATE " + quote + actualTable + quote + " SET " + quote + tempColumnName + quote + " = '' WHERE " + quote + tempColumnName + quote + " IS NULL";
+			sqls.add(replaceNullsSql);
+			final String setNotNullSql = "ALTER TABLE " + quote + actualTable + quote + " MODIFY " + quote + tempColumnName + quote + " NOT NULL";
+			sqls.add(setNotNullSql);
+		}
+		sqls.add("ALTER TABLE " + quote + actualTable + quote + " DROP COLUMN " + quote + columnName + quote);
+		final String renameSql = "ALTER TABLE " + quote + actualTable + quote + " RENAME COLUMN " + quote + tempColumnName + quote + " TO " + quote + columnName + quote;
+		sqls.add(renameSql);
+		for (final Map<String, Object> idx : depIndexes) 
+		{
+			final String indexName = (String) idx.get("name");
+			final List<String> cols = (List<String>) idx.get("columns");
+			final boolean unique = ((Boolean) idx.get("unique")).booleanValue();
+			sqls.add(getCreateIndexSQL(indexName, actualTable, cols, unique));
+		}
+		if(isPkColumn) 
+		{
+			boolean allNotNull = true;
+			for (final String pkCol : pkCols) 
+			{
+				String isNullableStr;
+				if(pkCol.equals(columnName.toUpperCase()))
+					isNullableStr = effectiveNewNullable ? "YES" : "NO";
+				else 
+				{
+					final ResultSet pkColRs = metaData.getColumns(null, null, actualTable, pkCol);
+					isNullableStr = pkColRs.next() ? pkColRs.getString("IS_NULLABLE") : "YES";
+					pkColRs.close();
+				}
+				if("YES".equals(isNullableStr)) 
+				{
+					allNotNull = false;
+					break;
+				}
+			}
+			Collections.sort(pkCols);
+			if(allNotNull)
+				sqls.add(getAddPrimaryKeySQL(actualTable, pkCols, null));
+			else
+				sqls.add(getCreateIndexSQL("UNQ_" + actualTable.replaceAll("[^A-Z0-9]","_"), actualTable, pkCols, true));
+		}
+		return sqls;
+	}
+
+	/**
+	 * Generates MODIFY COLUMN SQL with new size and nullability
+	 * Note: Modifying nullability or size may have restrictions in some DBs (e.g., can't reduce size if data exists)
+	 */
 	public String getModifyColumnSQL(final String tableName, final String columnName, final String newType, final Integer newSize, final Boolean newNullable) throws SQLException
 	{
 		final String quote=metaData.getIdentifierQuoteString();
 		String modifyClause;
-		if(productName.contains("mysql") || productName.contains("oracle"))
+		if(productName.contains("mysql"))
 			modifyClause=" MODIFY COLUMN ";
+		else
+		if(productName.contains("oracle"))
+			modifyClause=" MODIFY ";
 		else
 			modifyClause=" ALTER COLUMN ";
 		final StringBuilder sql=new StringBuilder("ALTER TABLE ").append(quote).append(tableName).append(quote)
@@ -943,14 +1123,18 @@ public class DDLGenerator
 	 */
 	private String getDataTypeSQL(final String type, final Integer size) throws SQLException
 	{
-		final String baseType=mapPortableType(type);
+		final String newType=mapPortableToLocalType(type);
+		if((productName.contains("oracle"))
+		&&(newType.contains("CHAR"))
+		&&(size == null))
+			throw new SQLException("Where's the size?!");
 		if(type.equalsIgnoreCase("VARCHAR"))
 		{
 			if(size!=null)
-				return baseType+"("+size+")";
+				return newType+"("+size+")";
 			throw new SQLException("Where's the size?!");
 		}
-		return baseType;
+		return newType;
 	}
 
 	/**
@@ -960,8 +1144,10 @@ public class DDLGenerator
 	 * @param portableType The portable type string
 	 * @return The mapped database-specific type string
 	 */
-	private String mapPortableType(final String portableType)
+	private String mapPortableToLocalType(final String portableType)
 	{
+		if(!typeMappings.containsKey(portableType))
+			System.out.println("STOP!");
 		return this.typeMappings.getOrDefault(portableType.toUpperCase(), portableType);
 	}
 
@@ -986,7 +1172,7 @@ public class DDLGenerator
 	 */
 	public String getDropColumnSQL(final String tableName, final String columnName) throws SQLException
 	{
-		if(!metaData.supportsAlterTableWithDropColumn())
+		if(!metaData.supportsAlterTableWithDropColumn() && !productName.contains("oracle"))
 			throw new SQLException("Drop column not supported");
 		final String quote=metaData.getIdentifierQuoteString();
 		return "ALTER TABLE "+quote+tableName+quote+" DROP COLUMN "+quote+columnName+quote+"";
@@ -1225,6 +1411,9 @@ public class DDLGenerator
 						if(productName.contains("postgres"))
 							sqls.addAll(getModifyColumnSQLPostgreSQL(actualTable, colName, desType, desSize, desNullable));
 						else
+						if(productName.contains("oracle"))
+							sqls.addAll(getModifyColumnSQLOracle(actualTable, colName, desType, desSize, desNullable));
+						else
 							sqls.add(getModifyColumnSQL(table, colName, desType, desSize, desNullable));
 					}
 				}
@@ -1251,6 +1440,9 @@ public class DDLGenerator
 				else
 				if(productName.contains("postgres"))
 					sqls.addAll(getModifyColumnSQLPostgreSQL(actualTable, col, type, size, nullable));
+				else
+				if(productName.contains("oracle"))
+					sqls.addAll(getModifyColumnSQLOracle(actualTable, col, type, size, nullable));
 				else
 					sqls.add(getModifyColumnSQL(table, col, type, size, nullable));
 			}
